Docker Hub specific changes in README
The three images are now also located on Docker Hub. Added a sections that describes the the pulling. Also changed the image tag names, since those had to be changed. "basalt" was already taken on Docker Hub. The usage example should be easier to understand with the $PWD, because the other path was just arbitrary.